Major modifications
The current modifications provide more details about the resource performance
as well make persistent the information relevant to Task and Resource.

For Task, new fields were added such as:

parameters
function
lastvisit
nextvisit
submittedtime

Reviewing them from bottom to top, the submittedtime is the new task identifier.
It contains the time when the task was submitted by the user and not represent
the time when the task was submitted to the grid.
It represents a time in milliseconds.

The nextvisit field is relevant for monitoring purposes and indicates when a
'GlobusjobstatusJob' instance will check the task status.

The lastvisit field says the last time when a 'GlobusjobstatusJob' instance
checked the task status.

'function' and 'parameters' fields were added for re-starting purposes, that means,
if a task instance achieves a 'FAILED' state, some external module might
re-start the task execution time later.

New additions
-------------

New service, job and domain files were added to the main development tree.
A 'MonitorService' was added. It retrieves information from remote resources through
the monitoring tools to exhibit resource information through XML and HTTP protocols.
A 'MonitorJob' periodically monitors the resources and updates the IS information when
it is necessary.
A 'Accountingresource' is a domain-class object to provide basic information about
how many tasks were submitted to a given resource, and how many of them were succeeded.
A 'Resourcecharacteristics' domain-class indicates the physical characteristics of the
resources.
In that sense, it provides information about number of nodes, number of cores per cpu,
cpu speed, and how many nodes are death.