Changes in the 'status' field and new way to determine the frequency which be used to query the task status.
The 'status' field in 'Accoutingresource' domain class changed from
'boolean' to 'int'. The possible values taken by the field are
especified in 'grails-app/domaing/Accountingresource.groovy'.

Finally, a new time gap between consecutive queries for the status of a
task has changed. Now, the frequency is not determined by observable
factors but values that the framework user defines in the 'application.properties'
file. Look to files attributes ending with 'THRESHOLD' word.