New attribute defined on 'Accoutingresource' and clearance procedure.
The 'tasksassigned' attribute records the number of tasks that were
assigned to a given gridresource when a task suffered an abnormal
finalization procedure.
This information can be used for determining what is the threshold
under the tasks on a given gridresource start to fail in achieve a
succeeded finalization state.

Since we want to test our framework with a high work load, we expect
that it cleans the processes and files to remain after a normal
or abnormal execution.