Altering the layout of the first parameter in 'executeTask' method.
There was altered the 'executeTask' method in two ways.
First, this method expects that the first parameter be a String divided
in three chunks by the symbol '|'. The first chunck corresponds to the
application name, the second is the execution name and the last parameter
corresponds to the application parameters.
Thereby, other 'LaunchService' execute  method is invoked.