Multiple modifications.
Due to 'komolongma' is experimenting problems with its perl script
that integrates SGE with GT4, it was removed from the list resources
defined in 'BootStrap' file.

A new attribue was added to 'Gridresource' class. This attribute
is just a sequence number. Despite of GORM defines a unique 'id'
to database level, it is not visible to model level, then a specific
sequence field has been added.

'RrschedService' was altered in order to look resources by sequence
instead of 'Gridresource' attribute since it was causing mismatching
types.