New classes for supporting the 'RrstoreschedService' class.
This service will enqueue all the received tasks which can not be
dispatched due to lack of available resources.

'Pendingtask' stores the application name, executable name, parameters,
arrival date/time of the task, and date/time when the task is attended.

'RrstoreschedService' has a similar behavior as 'RrschedService', however
it only will dispatch tasks where idle resources can be found otherwise
it will store the request in the 'Pendingtask' table.